Add MHC kinetic model as option
The original method for computing the overpotential relied on the
Butler-Volmer model of kinetics. Through this commit, we extend the
functionality to include the MHC model, which shows curvature in
the Tafel slope as a function of overpotential, leading to greater
overpotentials required for more moderate current densities applied.
An analysis of the applicability of MHC to CO2 reduction can be
found in Brown and Orella et al (2020 - in prep). For this purpose,
if a maximum current density exists below the requested current, the
applied current density is reset to 95% of the maximum to avoid over-
limiting currents during steady-state operation.